backend/requirements.txt
bcrypt==4.2.1
fastapi==0.115.6
uvicorn[standard]==0.32.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic[email]==2.10.3
python-multipart==0.0.20
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
